Add a quickstart section
In daily work I found that I frequently need to spin up the sever to test the client; the command to do it is in the docs, but it's not easy to locate quickly. This PR adds it close to the top of `README.md` to make it easy to find.